requirements.txt
Flask==2.2.3
SQLAlchemy==2.0.10
PyJWT~=2.6.0